Upgrade to Surefire 3.0.0-M6-SNAPSHOT
Now we can switch to SurefireForkNodeFactory, which enables Java agents
to dump information on stdOut and stdErr without creating dump files.
This way, we can simply read the (error) logs on the Maven console.

We do have to use a snapshot, because 3.0.0-M5 does not contain
necessary fixes which avoid the process from freezing and M6 has not
been released yet.